+++ tests/PUMS/local_loader.py
@@ -1,7 +1,7 @@
 from ingest.PUMS.PUMS_data import PUMSData
-from aggregate.count_PUMS_economics import PUMSCountEconomics
-from aggregate.count_PUMS_demographics import PUMSCountDemographics
-from aggregate.median_PUMS_demographics import PUMSMedianDemographics
+from aggregate.PUMS.count_PUMS_economics import PUMSCountEconomics
+from aggregate.PUMS.count_PUMS_demographics import PUMSCountDemographics
+from aggregate.PUMS.median_PUMS_demographics import PUMSMedianDemographics
 
 
 class LocalLoader:
